MDL-59911 core: skip on delete if there is no valid id

+++ files/converter/unoconv/classes/converter.php
@@ -194,7 +194,9 @@ class converter implements \core_files\converter_interface {
 
         $conversions = conversion::get_conversions_for_file($testdocx, $format);
         foreach ($conversions as $conversion) {
-            $conversion->delete();
+            if($conversion->get('id') < 0) {
+                $conversion->delete();
+            }
         }
 
         $conversion = new conversion(0, (object) [
